BCG Metric Membership

Biological Condition Gradient fuzzy membership for metrics.

BCG.Metric.Membership(
  df.metrics,
  df.rules,
  input.shape = "wide",
  col_SAMPLEID = "SAMPLEID",
  col_INDEX_NAME = "INDEX_NAME",
  col_INDEX_CLASS = "INDEX_CLASS",
  col_LEVEL = "LEVEL",
  col_METRIC_NAME = "METRIC_NAME",
  col_RULE_TYPE = "RULE_TYPE",
  col_LOWER = "LOWER",
  col_UPPER = "UPPER",
  col_METRIC_VALUE = "METRIC_VALUE",
  col_INCREASE = "INCREASE",
  ...
)

Arguments

df.metrics

Wide data frame with metric values to be evaluated.

df.rules

Data frame of metric thresholds to check.

input.shape

Shape of df.metrics; wide or long. Default is wide.

col_SAMPLEID

Column name for sample id. Default = "SAMPLEID"

col_INDEX_NAME

Column name for index name. Default = "INDEX_NAME"

col_INDEX_CLASS

Column name for index class Default = "INDEX_CLASS"

col_LEVEL

Column name for level. Default = "LEVEL"

col_METRIC_NAME

Column name for metric name. Default = "METRIC_NAME"

col_RULE_TYPE

Column name for rule type (e.g., Rule0). Default = "RULE_TYPE"

col_LOWER

Column name for lower limit. Default = "LOWER"

col_UPPER

Column name for upper limit. Default = "UPPER"

col_METRIC_VALUE

Column name for metric value. Default = "METRIC_VALUE"

col_INCREASE

Column name for if the metric value increases. Default = "INCREASE"

...

Arguments passed to `BCG.MetricMembership` used internally

Value

Returns a data frame of results in the long format.

Details

Converts metric values into BCG membership values. Uses a rules table to define the metrics, scoring range, and direction for each named index.

Deprecated col_SITE_TYPE for col_INDEX_CLASS in v2.0.0.9001.

Examples

# library(readxl)
# library(BioMonTools)

# Calculate Metrics
df_samps_bugs <- readxl::read_excel(
                           system.file("extdata/Data_BCG_PugLowWilVal.xlsx"
                                             , package = "BCGcalc")
                           , guess_max = 10^6)
myDF <- df_samps_bugs
myCols <- c("Area_mi2", "SurfaceArea", "Density_m2", "Density_ft2")
# populate missing columns prior to metric calculation
col_missing <- c("INFRAORDER", "HABITAT", "ELEVATION_ATTR", "GRADIENT_ATTR"
                 , "WSAREA_ATTR", "HABSTRUCT", "UFC")
myDF[, col_missing] <- NA
df_met_val_bugs <- BioMonTools::metric.values(myDF
                                              , "bugs"
                                              , fun.cols2keep = myCols)
#> Joining with `by = join_by(SAMPLEID, INDEX_NAME, INDEX_CLASS)`

# Import Rules
df_rules <- readxl::read_excel(system.file("extdata/Rules.xlsx"
                                           , package = "BCGcalc")
                      , sheet="Rules") 
#> Warning: Expecting numeric in C2070 / R2070C3: got 'Level 2'
#> Warning: Expecting numeric in C2071 / R2071C3: got 'Level 2'
#> Warning: Expecting numeric in C2072 / R2072C3: got 'Level 2'
#> Warning: Expecting numeric in C2073 / R2073C3: got 'Level 2'
#> Warning: Expecting numeric in C2074 / R2074C3: got 'Level 2'
#> Warning: Expecting numeric in C2075 / R2075C3: got 'Level 2'
#> Warning: Expecting numeric in C2076 / R2076C3: got 'Level 2'
#> Warning: Expecting numeric in C2077 / R2077C3: got 'Level 2'
#> Warning: Expecting numeric in C2078 / R2078C3: got 'Level 2'
#> Warning: Expecting numeric in C2079 / R2079C3: got 'Level 2'
#> Warning: Expecting numeric in C2080 / R2080C3: got 'Level 3'
#> Warning: Expecting numeric in C2081 / R2081C3: got 'Level 3'
#> Warning: Expecting numeric in C2082 / R2082C3: got 'Level 3'
#> Warning: Expecting numeric in C2083 / R2083C3: got 'Level 3'
#> Warning: Expecting numeric in C2084 / R2084C3: got 'Level 3'
#> Warning: Expecting numeric in C2085 / R2085C3: got 'Level 3'
#> Warning: Expecting numeric in C2086 / R2086C3: got 'Level 3'
#> Warning: Expecting numeric in C2087 / R2087C3: got 'Level 3'
#> Warning: Expecting numeric in C2088 / R2088C3: got 'Level 4'
#> Warning: Expecting numeric in C2089 / R2089C3: got 'Level 4'
#> Warning: Expecting numeric in C2090 / R2090C3: got 'Level 4'
#> Warning: Expecting numeric in C2091 / R2091C3: got 'Level 4'
#> Warning: Expecting numeric in C2092 / R2092C3: got 'Level 4'
#> Warning: Expecting numeric in C2093 / R2093C3: got 'Level 4'
#> Warning: Expecting numeric in C2094 / R2094C3: got 'Level 4'
#> Warning: Expecting numeric in C2095 / R2095C3: got 'Level 5'
#> Warning: Expecting numeric in C2096 / R2096C3: got 'Level 5'
#> Warning: Expecting numeric in C2097 / R2097C3: got 'Level 5'
#> Warning: Expecting numeric in C2098 / R2098C3: got 'Level 2'
#> Warning: Expecting numeric in C2099 / R2099C3: got 'Level 2'
#> Warning: Expecting numeric in C2100 / R2100C3: got 'Level 2'
#> Warning: Expecting numeric in C2101 / R2101C3: got 'Level 2'
#> Warning: Expecting numeric in C2102 / R2102C3: got 'Level 2'
#> Warning: Expecting numeric in C2103 / R2103C3: got 'Level 2'
#> Warning: Expecting numeric in C2104 / R2104C3: got 'Level 2'
#> Warning: Expecting numeric in C2105 / R2105C3: got 'Level 2'
#> Warning: Expecting numeric in C2106 / R2106C3: got 'Level 2'
#> Warning: Expecting numeric in C2107 / R2107C3: got 'Level 2'
#> Warning: Expecting numeric in C2108 / R2108C3: got 'Level 3'
#> Warning: Expecting numeric in C2109 / R2109C3: got 'Level 3'
#> Warning: Expecting numeric in C2110 / R2110C3: got 'Level 3'
#> Warning: Expecting numeric in C2111 / R2111C3: got 'Level 3'
#> Warning: Expecting numeric in C2112 / R2112C3: got 'Level 3'
#> Warning: Expecting numeric in C2113 / R2113C3: got 'Level 3'
#> Warning: Expecting numeric in C2114 / R2114C3: got 'Level 3'
#> Warning: Expecting numeric in C2115 / R2115C3: got 'Level 3'
#> Warning: Expecting numeric in C2116 / R2116C3: got 'Level 4'
#> Warning: Expecting numeric in C2117 / R2117C3: got 'Level 4'
#> Warning: Expecting numeric in C2118 / R2118C3: got 'Level 4'
#> Warning: Expecting numeric in C2119 / R2119C3: got 'Level 4'
#> Warning: Expecting numeric in C2120 / R2120C3: got 'Level 4'
#> Warning: Expecting numeric in C2121 / R2121C3: got 'Level 4'
#> Warning: Expecting numeric in C2122 / R2122C3: got 'Level 4'
#> Warning: Expecting numeric in C2123 / R2123C3: got 'Level 5'
#> Warning: Expecting numeric in C2124 / R2124C3: got 'Level 5'
#> Warning: Expecting numeric in C2125 / R2125C3: got 'Level 5'
#> Warning: Expecting numeric in C2126 / R2126C3: got 'Level 2'
#> Warning: Expecting numeric in C2127 / R2127C3: got 'Level 2'
#> Warning: Expecting numeric in C2128 / R2128C3: got 'Level 2'
#> Warning: Expecting numeric in C2129 / R2129C3: got 'Level 2'
#> Warning: Expecting numeric in C2130 / R2130C3: got 'Level 2'
#> Warning: Expecting numeric in C2131 / R2131C3: got 'Level 2'
#> Warning: Expecting numeric in C2132 / R2132C3: got 'Level 2'
#> Warning: Expecting numeric in C2133 / R2133C3: got 'Level 2'
#> Warning: Expecting numeric in C2134 / R2134C3: got 'Level 2'
#> Warning: Expecting numeric in C2135 / R2135C3: got 'Level 2'
#> Warning: Expecting numeric in C2136 / R2136C3: got 'Level 3'
#> Warning: Expecting numeric in C2137 / R2137C3: got 'Level 3'
#> Warning: Expecting numeric in C2138 / R2138C3: got 'Level 3'
#> Warning: Expecting numeric in C2139 / R2139C3: got 'Level 3'
#> Warning: Expecting numeric in C2140 / R2140C3: got 'Level 3'
#> Warning: Expecting numeric in C2141 / R2141C3: got 'Level 3'
#> Warning: Expecting numeric in C2142 / R2142C3: got 'Level 3'
#> Warning: Expecting numeric in C2143 / R2143C3: got 'Level 3'
#> Warning: Expecting numeric in C2144 / R2144C3: got 'Level 3'
#> Warning: Expecting numeric in C2145 / R2145C3: got 'Level 4'
#> Warning: Expecting numeric in C2146 / R2146C3: got 'Level 4'
#> Warning: Expecting numeric in C2147 / R2147C3: got 'Level 4'
#> Warning: Expecting numeric in C2148 / R2148C3: got 'Level 4'
#> Warning: Expecting numeric in C2149 / R2149C3: got 'Level 4'
#> Warning: Expecting numeric in C2150 / R2150C3: got 'Level 4'
#> Warning: Expecting numeric in C2151 / R2151C3: got 'Level 4'
#> Warning: Expecting numeric in C2152 / R2152C3: got 'Level 4'
#> Warning: Expecting numeric in C2153 / R2153C3: got 'Level 5'
#> Warning: Expecting numeric in C2154 / R2154C3: got 'Level 5'
#> Warning: Expecting numeric in C2155 / R2155C3: got 'Level 5'
#> Warning: Expecting numeric in C2156 / R2156C3: got 'Level 2'
#> Warning: Expecting numeric in C2157 / R2157C3: got 'Level 2'
#> Warning: Expecting numeric in C2158 / R2158C3: got 'Level 2'
#> Warning: Expecting numeric in C2159 / R2159C3: got 'Level 2'
#> Warning: Expecting numeric in C2160 / R2160C3: got 'Level 2'
#> Warning: Expecting numeric in C2161 / R2161C3: got 'Level 2'
#> Warning: Expecting numeric in C2162 / R2162C3: got 'Level 2'
#> Warning: Expecting numeric in C2163 / R2163C3: got 'Level 2'
#> Warning: Expecting numeric in C2164 / R2164C3: got 'Level 2'
#> Warning: Expecting numeric in C2165 / R2165C3: got 'Level 3'
#> Warning: Expecting numeric in C2166 / R2166C3: got 'Level 3'
#> Warning: Expecting numeric in C2167 / R2167C3: got 'Level 3'
#> Warning: Expecting numeric in C2168 / R2168C3: got 'Level 3'
#> Warning: Expecting numeric in C2169 / R2169C3: got 'Level 3'
#> Warning: Expecting numeric in C2170 / R2170C3: got 'Level 3'
#> Warning: Expecting numeric in C2171 / R2171C3: got 'Level 3'
#> Warning: Expecting numeric in C2172 / R2172C3: got 'Level 4'
#> Warning: Expecting numeric in C2173 / R2173C3: got 'Level 4'
#> Warning: Expecting numeric in C2174 / R2174C3: got 'Level 4'
#> Warning: Expecting numeric in C2175 / R2175C3: got 'Level 5'
#> Warning: Expecting numeric in C2176 / R2176C3: got 'Level 2'
#> Warning: Expecting numeric in C2177 / R2177C3: got 'Level 2'
#> Warning: Expecting numeric in C2178 / R2178C3: got 'Level 2'
#> Warning: Expecting numeric in C2179 / R2179C3: got 'Level 2'
#> Warning: Expecting numeric in C2180 / R2180C3: got 'Level 2'
#> Warning: Expecting numeric in C2181 / R2181C3: got 'Level 2'
#> Warning: Expecting numeric in C2182 / R2182C3: got 'Level 2'
#> Warning: Expecting numeric in C2183 / R2183C3: got 'Level 2'
#> Warning: Expecting numeric in C2184 / R2184C3: got 'Level 3'
#> Warning: Expecting numeric in C2185 / R2185C3: got 'Level 3'
#> Warning: Expecting numeric in C2186 / R2186C3: got 'Level 3'
#> Warning: Expecting numeric in C2187 / R2187C3: got 'Level 3'
#> Warning: Expecting numeric in C2188 / R2188C3: got 'Level 3'
#> Warning: Expecting numeric in C2189 / R2189C3: got 'Level 4'
#> Warning: Expecting numeric in C2190 / R2190C3: got 'Level 4'
#> Warning: Expecting numeric in C2191 / R2191C3: got 'Level 4'
#> Warning: Expecting numeric in C2192 / R2192C3: got 'Level 5'

# Run function
df_met_memb <- BCG.Metric.Membership(df_met_val_bugs, df_rules)

# Show Results
#View(df_met_memb)

# Save Results
write.table(df_met_memb
            , file.path(tempdir(), "Metric_Membership.tsv")
            , row.names = FALSE
            , col.names = TRUE
            , sep = "\t")